api_v1/serializers/main/rates.py
from rest_framework import serializers
from main.models import EstablishmentRates

class EstablishmentRatesSerializer(serializers.ModelSerializer):
    class Meta:
        model = EstablishmentRates
        fields = (
            'name',
        )