requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
black==23.3.0
celery==5.3.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.2
defusedxml==0.7.1
Django==4.2.2
django-crum
djangorestframework-simplejwt
djoser
django-minio-storage
django-storages
django-cors-headers==3.13.0
django-filter==22.1
django-allauth==0.57.0
django-debug-toolbar==4.1.0
django-phonenumber-field
phonenumbers
django-environ==0.10.0
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
drf-spectacular==0.27.1
gunicorn==20.1.0
idna==3.6
inflection==0.5.1
isort==5.12.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
